Update README.md
My father was killed during heart surgery last year. I had a 1969 circular worth 76,000,000 some how 4 10k payments was added and it was said to be fro raping my daughter. I messaged president Biden and it was never re issues for the proper amount. I can't even turn it in I can't afford it and I don't know the code to do it free a developer already stole it and the police didn't do anything. Can anyone do anything to help for 15% of it? Otherwise I'm just gonna kill everyone because that was to raise my kids and it's ruined 7 years of my life. I went to jail and everything else was stolen from me and I lost my job and house. 3 vehicles gone.